+++ app.json
@@ -1,4 +1,4 @@
 {
   "name": "DappFace",
-  "displayName": "DappFace"
+  "displayName": "DAPPFACE"
 }
