+++ ios/DappFace/AppDelegate.m
@@ -30,7 +30,7 @@
   rootViewController.view = rootView;
   self.window.rootViewController = rootViewController;
   [self.window makeKeyAndVisible];
-  [RNSplashScreen show];
+//  [RNSplashScreen show];
   return YES;
 }
