+++ package.json
@@ -64,6 +64,7 @@
     "react-native-keychain": "^3.1.2",
     "react-native-material-ripple": "^0.9.0",
     "react-native-navigation": "^3.2.0",
+    "react-native-navigation-stack": "^0.1.2",
     "react-native-orientation": "^3.1.3",
     "react-native-os": "^1.0.1",
     "react-native-paper": "^2.16.0",
@@ -76,6 +77,9 @@
     "react-native-vector-icons": "^6.4.2",
     "react-native-webview": "^7.0.5",
     "react-navigation": "^4.0.10",
+    "react-navigation-drawer": "^2.2.2",
+    "react-navigation-stack": "^1.9.3",
+    "react-navigation-tabs": "^2.5.5",
     "react-redux": "^7.1.1",
     "react-timer-mixin": "^0.13.4",
     "readable-stream": "^3.4.0",
