+++ src/components/screens/index.ts
@@ -1,0 +1,61 @@
+export * from './browser'
+export * from './links'
+export * from './settings'
+export * from './wallet'
+
+// export function registerScreens(): void {
+//   Navigation.registerComponent(Screen.BROWSER, wrapWithProvider(Browser))
+//   Navigation.registerComponent(
+//     Screen.SETTINGS.BACKUP.BASE,
+//     wrapWithProvider(Settings.Backup),
+//   )
+//   Navigation.registerComponent(
+//     Screen.SETTINGS.BACKUP.QUIZ,
+//     wrapWithProvider(Settings.Quiz),
+//   )
+//   Navigation.registerComponent(
+//     Screen.SETTINGS.BASE,
+//     wrapWithProvider(Settings.Settings),
+//   )
+//   Navigation.registerComponent(
+//     Screen.SETTINGS.CURRENCY,
+//     wrapWithProvider(Settings.Currency),
+//   )
+//   Navigation.registerComponent(
+//     Screen.SETTINGS.NETWORK,
+//     wrapWithProvider(Settings.Network),
+//   )
+//   Navigation.registerComponent(
+//     Screen.SETTINGS.SEARCH_ENGINE,
+//     wrapWithProvider(Settings.SearchEngine),
+//   )
+//   Navigation.registerComponent(
+//     Screen.WALLET.TOKEN_SEARCH,
+//     wrapWithProvider(Wallet.TokenSearch),
+//   )
+//   Navigation.registerComponent(
+//     Screen.WALLET.CONFIRM,
+//     wrapWithProvider(Wallet.Confirm),
+//   )
+//   Navigation.registerComponent(
+//     Screen.WALLET.IMPORT.ACCOUNT_SELECTOR,
+//     wrapWithProvider(Wallet.AccountSelector),
+//   )
+//   Navigation.registerComponent(
+//     Screen.WALLET.IMPORT.BASE,
+//     wrapWithProvider(Wallet.Import),
+//   )
+//   Navigation.registerComponent(
+//     Screen.WALLET.RECEIVE,
+//     wrapWithProvider(Wallet.Receive),
+//   )
+//   Navigation.registerComponent(
+//     Screen.WALLET.SCAN,
+//     wrapWithProvider(Wallet.Scan),
+//   )
+//   Navigation.registerComponent(
+//     Screen.WALLET.SEND,
+//     wrapWithProvider(Wallet.Send),
+//   )
+//   Navigation.registerComponent(Screen.LINKS, wrapWithProvider(Links))
+// }
