+++ src/components/screens/wallet/account-list/index.tsx
@@ -4,7 +4,6 @@ import {useSelector} from 'react-redux'
 import {Item} from 'src/components/screens/wallet/account-list/item'
 import {Size} from 'src/const'
 import {useDimensions} from 'src/hooks'
-import {pushReceive, pushSend, showBackup} from 'src/navigation'
 import {accountHook, accountSelector} from 'src/redux/module/account'
 import {entityType} from 'src/redux/module/entity'
 import {uiHook, uiType} from 'src/redux/module/ui'
